Extract 'Fonts' settings from 'Appearance' settings page
