+++ GitUI/CommandsDialogs/SettingsDialog/Pages/AppearanceSettingsPage.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Diagnostics;
-using System.Drawing;
 using System.IO;
 using System.Windows.Forms;
 using GitCommands;
@@ -15,10 +14,6 @@ namespace GitUI.CommandsDialogs.SettingsDialog.Pages
             new TranslationString("None");
         private readonly TranslationString _noDictFilesFound =
             new TranslationString("No dictionary files found in: {0}");
-
-        private Font _diffFont;
-        private Font _applicationFont;
-        private Font _commitFont;
         private readonly IImageCache _avatarCache;
 
         public AppearanceSettingsPage()
@@ -30,6 +25,11 @@ namespace GitUI.CommandsDialogs.SettingsDialog.Pages
             _avatarCache = new DirectoryImageCache(AppSettings.GravatarCachePath, AppSettings.AuthorImageCacheDays);
 
             NoImageService.Items.AddRange(Enum.GetNames(typeof(DefaultImageType)));
+        }
+
+        public static SettingsPageReference GetPageReference()
+        {
+            return new SettingsPageReferenceByType(typeof(AppearanceSettingsPage));
         }
 
         private static int GetTruncatePathMethodIndex(string text)
@@ -99,10 +99,6 @@ namespace GitUI.CommandsDialogs.SettingsDialog.Pages
             }
 
             chkShowRelativeDate.Checked = AppSettings.RelativeDate;
-
-            SetCurrentApplicationFont(AppSettings.Font);
-            SetCurrentDiffFont(AppSettings.DiffFont);
-            SetCurrentCommitFont(AppSettings.CommitFont);
         }
 
         protected override void PageToSettings()
@@ -122,10 +118,6 @@ namespace GitUI.CommandsDialogs.SettingsDialog.Pages
             AppSettings.RelativeDate = chkShowRelativeDate.Checked;
 
             AppSettings.Dictionary = Dictionary.SelectedIndex == 0 ? "none" : Dictionary.Text;
-
-            AppSettings.DiffFont = _diffFont;
-            AppSettings.Font = _applicationFont;
-            AppSettings.CommitFont = _commitFont;
         }
 
         private void Dictionary_DropDown(object sender, EventArgs e)
@@ -150,62 +142,6 @@ namespace GitUI.CommandsDialogs.SettingsDialog.Pages
             {
                 MessageBox.Show(this, string.Format(_noDictFilesFound.Text, AppSettings.GetDictionaryDir()));
             }
-        }
-
-        private void diffFontChangeButton_Click(object sender, EventArgs e)
-        {
-            diffFontDialog.Font = _diffFont;
-            DialogResult result = diffFontDialog.ShowDialog(this);
-
-            if (result == DialogResult.OK || result == DialogResult.Yes)
-            {
-                SetCurrentDiffFont(diffFontDialog.Font);
-            }
-        }
-
-        private void applicationFontChangeButton_Click(object sender, EventArgs e)
-        {
-            applicationDialog.Font = _applicationFont;
-            DialogResult result = applicationDialog.ShowDialog(this);
-
-            if (result == DialogResult.OK || result == DialogResult.Yes)
-            {
-                SetCurrentApplicationFont(applicationDialog.Font);
-            }
-        }
-
-        private void commitFontChangeButton_Click(object sender, EventArgs e)
-        {
-            commitFontDialog.Font = _commitFont;
-            DialogResult result = commitFontDialog.ShowDialog(this);
-
-            if (result == DialogResult.OK || result == DialogResult.Yes)
-            {
-                SetCurrentCommitFont(commitFontDialog.Font);
-            }
-        }
-
-        private void SetCurrentDiffFont(Font newFont)
-        {
-            _diffFont = newFont;
-            SetFontButtonText(newFont, diffFontChangeButton);
-        }
-
-        private void SetCurrentApplicationFont(Font newFont)
-        {
-            _applicationFont = newFont;
-            SetFontButtonText(newFont, applicationFontChangeButton);
-        }
-
-        private void SetCurrentCommitFont(Font newFont)
-        {
-            _commitFont = newFont;
-            SetFontButtonText(newFont, commitFontChangeButton);
-        }
-
-        private static void SetFontButtonText(Font font, Button button)
-        {
-            button.Text = string.Format("{0}, {1}", font.FontFamily.Name, (int)(font.Size + 0.5f));
         }
 
         private void ClearImageCache_Click(object sender, EventArgs e)
