+++ GitUI/GitUI.csproj
@@ -219,6 +219,12 @@
     <Compile Include="CommandsDialogs\SettingsDialog\FormAvailableEncodings.Designer.cs">
       <DependentUpon>FormAvailableEncodings.cs</DependentUpon>
     </Compile>
+    <Compile Include="CommandsDialogs\SettingsDialog\Pages\AppearanceFontsSettingsPage.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="CommandsDialogs\SettingsDialog\Pages\AppearanceFontsSettingsPage.Designer.cs">
+      <DependentUpon>AppearanceFontsSettingsPage.cs</DependentUpon>
+    </Compile>
     <Compile Include="CommandsDialogs\SettingsDialog\Pages\DiffViewerSettingsPage.cs">
       <SubType>UserControl</SubType>
     </Compile>
@@ -1175,6 +1181,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="CommandsDialogs\SettingsDialog\FormAvailableEncodings.resx">
       <DependentUpon>FormAvailableEncodings.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="CommandsDialogs\SettingsDialog\Pages\AppearanceFontsSettingsPage.resx">
+      <DependentUpon>AppearanceFontsSettingsPage.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="CommandsDialogs\SettingsDialog\Pages\DiffViewerSettingsPage.resx">
       <DependentUpon>DiffViewerSettingsPage.cs</DependentUpon>
